requirements.txt
python-telegram-bot==13.7
requests==2.28.2
openai==1.5.0
supabase==1.0.3
python-dotenv==1.0.0